mds: issue caps on file update finish if no client cap specified
Fix the following scenario:

client released caps on one inode. Then, the inode's ifile was wrlocked
during updating client range. Before the update had finished, client
re-opened the file again for reading. Since ifile was wrlocked, the client
was not given the Fr cap and was blocked on reading forever.